Some fixes: - client first reports are now mapped correctly - symptom ids are now stored correctly - Id of diaryentries is returned correctly

+++ src/main/java/de/wevsvirushackathon/coronareport/diary/DiaryEntryDtoIn.java
@@ -22,7 +22,7 @@ public class DiaryEntryDtoIn {
     private float bodyTemperature;
     
     @Getter @Setter
-    private List<Long> symptoms;
+    private List<Integer> symptoms;
     
     @Getter @Setter
     private boolean transmittedToHealthDepartment;
